Move commons package to Model package
