+++ src/main/java/seedu/address/logic/commands/AddBoxCommand.java
@@ -18,10 +18,10 @@ import seedu.address.model.person.Box;
 import seedu.address.model.person.DeliveryStatus;
 import seedu.address.model.person.Email;
 import seedu.address.model.person.ExpiryDate;
-import seedu.address.commons.core.name.Name;
 import seedu.address.model.person.Person;
-import seedu.address.commons.core.phone.Phone;
 import seedu.address.model.person.Remark;
+import seedu.address.model.commons.name.Name;
+import seedu.address.model.commons.phone.Phone;
 import seedu.address.model.tag.Tag;
 
 /**
