+++ src/main/java/seedu/address/logic/parser/AddCommandParser.java
@@ -19,10 +19,10 @@ import seedu.address.model.person.Address;
 import seedu.address.model.person.Box;
 import seedu.address.model.person.Email;
 import seedu.address.model.person.ExpiryDate;
-import seedu.address.commons.core.name.Name;
 import seedu.address.model.person.Person;
-import seedu.address.commons.core.phone.Phone;
 import seedu.address.model.person.Remark;
+import seedu.address.model.commons.name.Name;
+import seedu.address.model.commons.phone.Phone;
 import seedu.address.model.tag.Tag;
 
 /**
