+++ src/main/java/seedu/address/logic/parser/DeleteBoxCommandParser.java
@@ -11,7 +11,7 @@ import java.util.stream.Stream;
 
 import seedu.address.logic.commands.DeleteBoxCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.commons.core.name.Name;
+import seedu.address.model.commons.name.Name;
 
 /**
  * Parses input arguments and creates a new DeleteBoxCommand object
