+++ src/main/java/seedu/address/model/commons/phone/Phone.java
@@ -1,4 +1,4 @@
-package seedu.address.commons.core.phone;
+package seedu.address.model.commons.phone;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.AppUtil.checkArgument;
