+++ src/main/java/seedu/address/model/delivery/Driver.java
@@ -5,8 +5,8 @@ import static seedu.address.commons.util.CollectionUtil.requireAllNonNull;
 import java.util.Objects;
 
 import seedu.address.commons.util.ToStringBuilder;
-import seedu.address.commons.core.name.Name;
-import seedu.address.commons.core.phone.Phone;
+import seedu.address.model.commons.name.Name;
+import seedu.address.model.commons.phone.Phone;
 
 /**
  * Represents a Driver in Client2Door.
