+++ src/main/java/seedu/address/model/person/Person.java
@@ -8,9 +8,9 @@ import java.util.Objects;
 import java.util.Set;
 import java.util.TreeSet;
 
-import seedu.address.commons.core.name.Name;
-import seedu.address.commons.core.phone.Phone;
 import seedu.address.commons.util.ToStringBuilder;
+import seedu.address.model.commons.name.Name;
+import seedu.address.model.commons.phone.Phone;
 import seedu.address.model.tag.Tag;
 
 /**
