+++ src/test/java/seedu/address/logic/commands/AddBoxCommandTest.java
@@ -15,8 +15,8 @@ import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
 import seedu.address.model.person.Box;
 import seedu.address.model.person.ExpiryDate;
-import seedu.address.commons.core.name.Name;
 import seedu.address.model.person.Person;
+import seedu.address.model.commons.name.Name;
 
 public class AddBoxCommandTest {
