+++ src/test/java/seedu/address/logic/commands/DeleteBoxCommandTest.java
@@ -18,8 +18,8 @@ import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
 import seedu.address.model.person.Box;
-import seedu.address.commons.core.name.Name;
 import seedu.address.model.person.Person;
+import seedu.address.model.commons.name.Name;
 
 public class DeleteBoxCommandTest {
