+++ src/test/java/seedu/address/logic/parser/AddBoxCommandParserTest.java
@@ -12,7 +12,7 @@ import org.junit.jupiter.api.Test;
 import seedu.address.logic.commands.AddBoxCommand;
 import seedu.address.model.person.Box;
 import seedu.address.model.person.ExpiryDate;
-import seedu.address.commons.core.name.Name;
+import seedu.address.model.commons.name.Name;
 
 public class AddBoxCommandParserTest {
