+++ src/test/java/seedu/address/logic/parser/EditBoxCommandParserTest.java
@@ -16,7 +16,7 @@ import seedu.address.logic.commands.EditBoxCommand;
 import seedu.address.logic.commands.EditBoxCommand.EditBoxDescriptor;
 import seedu.address.model.person.Box;
 import seedu.address.model.person.ExpiryDate;
-import seedu.address.commons.core.name.Name;
+import seedu.address.model.commons.name.Name;
 import seedu.address.testutil.EditBoxDescriptorBuilder;
 
 public class EditBoxCommandParserTest {
