+++ src/test/java/seedu/address/model/commons/PhoneTest.java
@@ -1,11 +1,12 @@
-package seedu.address.model.person;
+package seedu.address.model.commons;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static seedu.address.testutil.Assert.assertThrows;
 
 import org.junit.jupiter.api.Test;
-import seedu.address.commons.core.phone.Phone;
+
+import seedu.address.model.commons.phone.Phone;
 
 public class PhoneTest {
